+++ src/main/java/com/sigma/cinematicketpro/service/AuthService.java
@@ -1,7 +1,7 @@
 package com.sigma.cinematicketpro.service;
 
-import com.sigma.cinematicketpro.dto.AuthenticationRequest;
-import com.sigma.cinematicketpro.dto.RegistrationRequest;
+import com.sigma.cinematicketpro.dto.auth.AuthenticationRequest;
+import com.sigma.cinematicketpro.dto.auth.RegistrationRequest;
 
 public interface AuthService {
     String register(RegistrationRequest registrationRequest);
