+++ src/test/java/com/sigma/cinematicketpro/service/impl/AuthServiceImplTest.java
@@ -1,8 +1,8 @@
 package com.sigma.cinematicketpro.service.impl;
 
 import com.sigma.cinematicketpro.TestUtils;
-import com.sigma.cinematicketpro.dto.AuthenticationRequest;
-import com.sigma.cinematicketpro.dto.RegistrationRequest;
+import com.sigma.cinematicketpro.dto.auth.AuthenticationRequest;
+import com.sigma.cinematicketpro.dto.auth.RegistrationRequest;
 import com.sigma.cinematicketpro.entity.CtpUser;
 import com.sigma.cinematicketpro.entity.Role;
 import com.sigma.cinematicketpro.exception.UserAlreadyExistsException;
